Add files via upload
The task of this project is to predict whether a customer will purchase a product or service based on their features and activities. There are total 4 data sets which include customer details about their Job, Marital, Loan, Balance, Default, Housing and other financial details. As there are data sets with different features Models are concluded separately but aim to predict result combined.